+++ CHANGELOG.md
@@ -24,6 +24,22 @@ Language
   The reduction machine has been rewritten from scratch and should be faster
   than the old one in all cases, even those not exploiting laziness.
 
+* Compile-time inlining.
+
+  Simple definitions (that don't do any pattern matching) marked as INLINE are
+  now also inlined at compile time, whereas before they were only inlined by
+  the compiler backends. Inlining only triggers in function bodies and not in
+  type signatures, to preserve goal types as far as possible.
+
+* Automatic inlining.
+
+  Definitions satisfying the following criteria are now automatically inlined
+  (can be disabled using the new NOINLINE pragma):
+
+    - No pattern matching.
+    - Uses each argument at most once.
+    - Does not use all its arguments.
+
 ### Syntax
 
 * Do-notation.
@@ -358,6 +374,14 @@ Pragmas and options
 * The command line option `--help` now takes an optional argument which
   allows the user to request more specific usage information about particular
   topics. The only one added so far is `warning`.
+
+* New pragma NOINLINE.
+
+  ```agda
+  {-# NOINLINE f #-}
+  ```
+
+  Disables automatic inlining of `f`.
 
 Emacs mode
 ----------